requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
black==24.10.0
cachetools==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
flake8==7.1.1
frozenlist==1.4.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.7
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pydantic==2.7.3
pydantic_core==2.18.4
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
